pkg: Move MetricFamilyDef to metrics pkg
